+++ api/src/lib.rs
@@ -4,10 +4,19 @@
     box_pointers,
     future_incompatible,
     missing_copy_implementations,
+    // missing_docs,
     nonstandard_style,
     rust_2018_idioms,
     trivial_casts,
-    trivial_numeric_casts
+    trivial_numeric_casts,
+    unused_qualifications,
+    clippy::all,
+    clippy::complexity,
+    clippy::correctness,
+    clippy::pedantic,
+    clippy::perf,
+    clippy::nursery,
+    clippy::style
 )]
 
 #[cfg(test)]