+++ api/src/parser/term/annotation_composable.rs
@@ -11,8 +11,11 @@ impl<'a> AnnotationComposable for &'a [Term] {
     fn annotation_composition(self) -> Option<AnnotationComposition> {
         let mut map = self
             .iter()
-            .filter(|term| term.annotation.is_some())
-            .map(|term| (term.annotation.clone().unwrap(), term.exponent.unwrap_or(1)))
+            .filter_map(|term| {
+                term.annotation
+                    .as_ref()
+                    .map(|annotation| (annotation.clone(), term.exponent.unwrap_or(1)))
+            })
             .fold(AnnotationComposition::new(), |mut map, (key, exponent)| {
                 map.entry(key)
                     .and_modify(|entry| *entry += exponent)