+++ atom_generator/src/toml_structs/toml_custom_atom_list.rs
@@ -14,8 +14,6 @@ impl FromStr for TomlCustomAtomList {
     type Err = ::toml::de::Error;
 
     fn from_str(toml: &str) -> Result<Self, Self::Err> {
-        let toml_atom_list: TomlCustomAtomList = ::toml::from_str(toml)?;
-
-        Ok(toml_atom_list)
+        ::toml::from_str(toml)
     }
 }